Added path lookup functions
- Renamed BLOCK_SIZE and INODE_SIZE to SIZE since each is scoped by its type anyway.
- Created a DirectoryRecord type to cast entrines in a directory data block.
