+++ src/lib/Block.h
@@ -3,8 +3,8 @@
 #include <cstdint>
 
 struct Block {
-  static const uint16_t BLOCK_SIZE = 4096;
+  static const uint16_t SIZE = 4096;
   typedef uint64_t ID;
 
-  char data[BLOCK_SIZE];
+  char data[SIZE];
 };
