+++ src/lib/DirectoryRecord.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <cstdint>
+#include "INode.h"
+
+struct DirectoryRecord {
+	uint32_t length;
+	INode::ID inode_ID;
+	char name[];
+};
